fix index method error
